+++ source/unit/onStop.lua
@@ -1,0 +1,3 @@
+if databank ~= nil then
+    databank.setStringValue("options", json.encode(options))
+end
